+++ examples/minimal/gatsby-config.js
@@ -4,6 +4,10 @@ module.exports = {
       resolve: 'gatsby-plugin-react-redux',
       options: {
         pathToCreateStoreModule: './src/state/createStore',
+        cleanupOnClient: false,
+        serialize: {
+          isJSON: false,
+        },
       },
     },
   ],